+++ Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 
-[assembly: AssemblyTitle("PcNinja's Ultimate Office Installer")]
+[assembly: AssemblyTitle("PcNinja's Smart Office Installer")]
 [assembly: AssemblyDescription("Volume Deployment Tool by PcNinja.Pro - www.pcninja.pro")]
 [assembly: AssemblyCompany("PcNinja.Pro")]
-[assembly: AssemblyProduct("PcNinja Ultimate Office Installer")]
+[assembly: AssemblyProduct("PcNinja Smart Office Installer")]
 [assembly: AssemblyCopyright("(c) 2025 PcNinja.Pro - www.pcninja.pro")]
-[assembly: AssemblyFileVersion("3.3.4.2")]
+[assembly: AssemblyFileVersion("3.3.4.3")]
 [assembly: ComVisible(false)]
 [assembly: AssemblyVersion("3.3.0.0")]